Convert "Events" pages to typescript
